Move shared validate hooks to new command
